Add AnnotateItPermissions plugin, implementing permissions for AnnotateIt's API
After some discussion, I've agreed with @aron to commit this as a separate
subclass of the Permissions plugin, leaving the Permissions plugin as a
simpler example of using the annotation permissions field, while putting all
the rather more complex authorization logic used by AnnotateIt into its own
easy-to-use plugin.

In addition, I have cleaned up the implementation of the Permissions plugin,
leaving even more of the work to the userAuthorize function. Unfortunately
this has changed the signature of the userAuthorize function, and it will
necessarily be slightly more complicated than it previously was, but
ultimately it provides all the same default logic, with the ability to
completely customize the authorization behaviour.