annotateitpermissions: set user correctly for new annotations
The AnnotateItPermissions plugin previously partially inherited the Permissions
plugin's `addFieldsToAnnotation` method, causing the 'user' field on new
annotations to be set, incorrectly, to the entire auth token rather than the
string user ID of the authenticated user.

The result of this was that the authenticated user didn't have 'admin'
permissions on brand new (unsaved) annotations, and so the "Allow anyone to
{view,edit} this annotation" checkboxes didn't appear in the UI.

This commit adds a test and a fix for this behaviour.